# uzum_api_gateway/api/docs/swagger.yaml
definitions:
...
  catalog_service.CreateProductReview:
    properties:
      customer_id:
        type: string
      product_id:
        type: string
      text:
        type: string
    type: object
  catalog_service.Empty:
    type: object
...
  catalog_service.GetAllProductReviewResponse:
    properties:
      Count:
        type: integer
      ProductReviews:
        items:
          $ref: '#/definitions/catalog_service.ProductReview'
        type: array
    type: object
...
  catalog_service.ProductReview:
    properties:
      created_at:
        type: string
      customer_id:
        type: string
      id:
        type: string
      order_id:
        type: string
      product_id:
        type: string
      rating:
        type: number
      text:
        type: string
    type: object
...
  catalog_service.UpdateProductReview:
    properties:
      customer_id:
        type: string
      id:
        type: string
      product_id:
        type: string
      rating:
        type: number
      text:
        type: string
    type: object
  models.ResponseError:
    properties:
      error: {}
    type: object
...
  user_service.CreateShop:
    properties:
      currency:
        type: string
      description_en:
        type: string
      description_ru:
        type: string
      description_uz:
        type: string
      location:
        type: string
      name_en:
        type: string
      name_ru:
        type: string
      name_uz:
        type: string
      payment_types:
        items:
          type: string
        type: array
      phone:
        type: string
      slug:
        type: string
    type: object
...
  user_service.Empty4:
    type: object
...
  user_service.GetListShopResponse:
    properties:
      count:
        type: integer
      shops:
        items:
          $ref: '#/definitions/user_service.Shop'
        type: array
    type: object
...
  user_service.Shop:
    properties:
      created_at:
        type: string
      currency:
        type: string
      deleted_at:
        type: string
      description_en:
        type: string
      description_ru:
        type: string
      description_uz:
        type: string
      id:
        type: string
      location:
        type: string
      name_en:
        type: string
      name_ru:
        type: string
      name_uz:
        type: string
      payment_types:
        items:
          type: string
        type: array
      phone:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  user_service.UpdateShop:
    properties:
      currency:
        type: string
      description_en:
        type: string
      description_ru:
        type: string
      description_uz:
        type: string
      id:
        type: string
      location:
        type: string
      name_en:
        type: string
      name_ru:
        type: string
      name_uz:
        type: string
      payment_types:
        items:
          type: string
        type: array
      phone:
        type: string
      slug:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /CreateShop:
    post:
      consumes:
      - application/json
      description: API for creating shop
      parameters:
      - description: shop
        in: body
        name: shop
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateShop'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Shop'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create shop
      tags:
      - shop
...
  /DeleteShop/{id}:
    delete:
      consumes:
      - application/json
      description: API for deleting a shop by ID
      parameters:
      - description: shop ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Empty4'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete a shop by ID
      tags:
      - shop
...
  /GetByIdShop/{id}:
    get:
      consumes:
      - application/json
      description: API for getting a single shop by ID
      parameters:
      - description: shop ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Shop'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a single shop by ID
      tags:
      - shop
...
  /GetListShop:
    get:
      consumes:
      - application/json
      description: API for getting list of shops
      parameters:
      - description: shops
        in: query
        name: shop
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.GetListShopResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get list of shops
      tags:
      - shop
...
  /ProductReview:
    post:
      consumes:
      - application/json
      description: API for creating a category review
      parameters:
      - description: category review
        in: body
        name: ProductReview
        required: true
        schema:
          $ref: '#/definitions/catalog_service.CreateProductReview'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog_service.ProductReview'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a category review
      tags:
      - category
  /ProductReview/{id}:
    delete:
      consumes:
      - application/json
      description: API for deleting a category review by ID
      parameters:
      - description: category review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog_service.Empty'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete a category review by ID
      tags:
      - category
    get:
      consumes:
      - application/json
      description: API for getting a single category review by ID
      parameters:
      - description: category review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog_service.ProductReview'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a single category review by ID
      tags:
      - category
    put:
      consumes:
      - application/json
      description: API for updating a category review by ID
      parameters:
      - description: category review ID
        in: path
        name: id
        required: true
        type: string
      - description: category review
        in: body
        name: ProductReview
        required: true
        schema:
          $ref: '#/definitions/catalog_service.UpdateProductReview'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog_service.ProductReview'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update a category review by ID
      tags:
      - category
...
  /UpdateShop/{id}:
    put:
      consumes:
      - application/json
      description: API for updating a shop by ID
      parameters:
      - description: shop ID
        in: path
        name: id
        required: true
        type: string
      - description: shop
        in: body
        name: shop
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateShop'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Shop'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update a shop by ID
      tags:
      - shop
...
  /getlistProductReview:
    get:
      consumes:
      - application/json
      description: API for getting list category reviews
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog_service.GetAllProductReviewResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get List Category Reviews
      tags:
      - category
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"